requirements.txt
PyGithub>=1.55
pyyaml>=6.0